fixing clang warnings

+++ kcache.hpp
@@ -49,7 +49,7 @@ extern "C" {
 */
 #ifndef LASVM_KERNEL_T_DEFINED
 #define LASVM_KERNEL_T_DEFINED
-typedef double (*lasvm_kernel_t)(size_t i, size_t j, void* closure);
+typedef double (*lasvm_kernel_t)(long i, long j, void* closure);
 #endif
 
 
@@ -80,25 +80,25 @@ void lasvm_kcache_destroy(lasvm_kcache_t *self);
    Argument <entries> indicates the maximum cache memory in bytes
    The default size is 256Mb.
 */
-void lasvm_kcache_set_maximum_size(lasvm_kcache_t *self, size_t entries);
+void lasvm_kcache_set_maximum_size(lasvm_kcache_t *self, long entries);
 
 /* --- lasvm_kcache_get_maximum_size
    Returns the maximum cache memory.
  */
-size_t lasvm_kcache_get_maximum_size(lasvm_kcache_t *self);
+long lasvm_kcache_get_maximum_size(lasvm_kcache_t *self);
 
 /* --- lasvm_kcache_get_current_size
    Returns the currently used cache memory.
    This can slighly exceed the value specified by 
    <lasvm_kcache_set_maximum_size>.
  */
-size_t lasvm_kcache_get_current_size(lasvm_kcache_t *self);
+long lasvm_kcache_get_current_size(lasvm_kcache_t *self);
 
 /* --- lasvm_kcache_query
    Returns the possibly cached value of the Gram matrix element (<i>,<j>).
    This function will not modify the cache geometry.
  */
-double lasvm_kcache_query(lasvm_kcache_t *self, size_t i, size_t j);
+double lasvm_kcache_query(lasvm_kcache_t *self, long i, long j);
 
 /* --- lasvm_kcache_query_row
    Returns the <len> first elements of row <i> of the Gram matrix.
@@ -108,19 +108,19 @@ double lasvm_kcache_query(lasvm_kcache_t *self, size_t i, size_t j);
    and vice-versa.
 */
 
-double *lasvm_kcache_query_row(lasvm_kcache_t *self, size_t i, size_t len);
+double *lasvm_kcache_query_row(lasvm_kcache_t *self, long i, long len);
 
 /* --- lasvm_kcache_status_row
    Returns the number of cached entries for row i.
 */
 
-size_t lasvm_kcache_status_row(lasvm_kcache_t *self, size_t i);
+long lasvm_kcache_status_row(lasvm_kcache_t *self, long i);
 
 /* --- lasvm_kcache_discard_row
    Indicates that we wont need row i in the near future.
 */
 
-void lasvm_kcache_discard_row(lasvm_kcache_t *self, size_t i);
+void lasvm_kcache_discard_row(lasvm_kcache_t *self, long i);
 
 
 /* --- lasvm_kcache_i2r
@@ -129,8 +129,8 @@ void lasvm_kcache_discard_row(lasvm_kcache_t *self, size_t i);
    the conversion table from example index to row position and vice-versa. 
 */
 
-size_t *lasvm_kcache_i2r(lasvm_kcache_t *self, size_t n);
-size_t *lasvm_kcache_r2i(lasvm_kcache_t *self, size_t n);
+long *lasvm_kcache_i2r(lasvm_kcache_t *self, long n);
+long *lasvm_kcache_r2i(lasvm_kcache_t *self, long n);
 
 
 /* --- lasvm_kcache_swap_rr
@@ -141,9 +141,9 @@ size_t *lasvm_kcache_r2i(lasvm_kcache_t *self, size_t n);
    or by indicating the example number (<i1>, <i2>).
 */
 
-void lasvm_kcache_swap_rr(lasvm_kcache_t *self, size_t r1, size_t r2);
-void lasvm_kcache_swap_ii(lasvm_kcache_t *self, size_t i1, size_t i2);
-void lasvm_kcache_swap_ri(lasvm_kcache_t *self, size_t r1, size_t i2);
+void lasvm_kcache_swap_rr(lasvm_kcache_t *self, long r1, long r2);
+void lasvm_kcache_swap_ii(lasvm_kcache_t *self, long i1, long i2);
+void lasvm_kcache_swap_ri(lasvm_kcache_t *self, long r1, long i2);
 
 
 #ifdef __cplusplus__
